# src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Telegram Drive API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local server"
    }
  ],
...
  "paths": {
...
    "/share": {
      "get": {
        "tags": [
          "Sharing"
        ],
        "summary": "List all share links created by the user",
        "parameters": [
          {
            "in": "header",
            "name": "x-session-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of share links",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "token": { "type": "string" },
                          "share_type": { "type": "string" },
                          "folder_id": { "type": "number", "nullable": true },
                          "message_id": { "type": "number", "nullable": true },
                          "created_at": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/share/create": {
      "post": {
        "tags": [
          "Sharing"
        ],
        "summary": "Create a share link",
        "parameters": [
          {
            "in": "header",
            "name": "x-session-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "share_type": {
                    "type": "string",
                    "enum": [
                      "file",
                      "folder"
                    ]
                  },
                  "folder_id": {
                    "type": "number"
                  },
                  "message_id": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Share link created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "url": "/api/share/d290f1ee-6c54-4b01-90e6-d701748f0851"
                }
              }
            }
          }
        }
      }
    },
    "/share/{token}": {
      "get": {
        "tags": [
          "Sharing"
        ],
        "summary": "Get metadata for a shared file or folder",
        "description": "Returns JSON information about the share, including file details and owner status.",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Share metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "share_type": { "type": "string" },
                    "owner_phone": { "type": "string" },
                    "created_at": { "type": "string" },
                    "is_owner_online": { "type": "boolean" },
                    "file": { "type": "object", "nullable": true },
                    "folder": { "type": "object", "nullable": true },
                    "file_count": { "type": "number", "nullable": true }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/share/{token}/zip": {
      "get": {
        "tags": [
          "Sharing"
        ],
        "summary": "Download shared folder as a ZIP file",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ZIP file stream",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  }
}